requirement.txt
tensorflow==2.7.0
opencv-python==4.5.3.56
numpy==1.18.5
opencv-contrib-python==4.5.2.54
argparse==1.4.0
Pillow==8.4.0
